Add XOR benchmarks
